restructure card components
